# docs/swagger.yaml
basePath: /
definitions:
...
  common.jsonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handles.BatchRenameReq:
    properties:
      rename_objects:
        items:
          properties:
            new_name:
              type: string
            src_name:
              type: string
          type: object
        type: array
      src_dir:
        type: string
    type: object
...
  handles.RecursiveMoveReq:
    properties:
      conflict_policy:
        type: string
      dst_dir:
        type: string
      src_dir:
        type: string
    type: object
  handles.RegexRenameReq:
    properties:
      new_name_regex:
        type: string
      src_dir:
        type: string
      src_name_regex:
        type: string
    type: object
...
info:
  contact:
    name: OpenList Team
    url: https://github.com/OpenListTeam/
  description: A file list program that supports multiple storage.
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://doc.oplist.org/terms
  title: OpenList
  version: latest
paths:
...
  /api/fs/batch_rename:
    post:
      consumes:
      - application/json
      description: Rename multiple files in a specified directory using a list of
        source and new names
      parameters:
      - description: Batch rename request
        in: body
        name: rename
        required: true
        schema:
          $ref: '#/definitions/handles.BatchRenameReq'
      produces:
      - application/json
      responses:
        "200":
          description: Batch rename successful
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Authorization: []
      summary: Batch Rename Files
      tags:
      - Filesystem
...
  /api/fs/recursive_move:
    post:
      consumes:
      - application/json
      description: Move files/folders recursively (preserves directory structure)
      parameters:
      - description: Recursive move request
        in: body
        name: move
        required: true
        schema:
          $ref: '#/definitions/handles.RecursiveMoveReq'
      produces:
      - application/json
      responses:
        "200":
          description: Recursive move successful
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Authorization: []
      summary: Recursive move
      tags:
      - Filesystem
  /api/fs/regex_rename:
    post:
      consumes:
      - application/json
      description: Rename files in a specified directory using regex patterns for
        source and new names
      parameters:
      - description: Regex-based rename request
        in: body
        name: rename
        required: true
        schema:
          $ref: '#/definitions/handles.RegexRenameReq'
      produces:
      - application/json
      responses:
        "200":
          description: Regex-based rename successful
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Authorization: []
      summary: Regex-based Rename Files
      tags:
      - Filesystem
...
securityDefinitions:
  Authorization:
    description: Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"